LGTM.com warning: Variable defined multiple times
This assignment to 'x' is unnecessary as it is redefined here before this value is used.
This assignment to 'y' is unnecessary as it is redefined here before this value is used.
This assignment to 'z' is unnecessary as it is redefined here before this value is used.

Also moved the `image_comments` parsing under `else:` because there is
no need to parse after explictly setting `image_comments = b"b=0(0,0,0)"`.